ml: Lab3, Lab4, cf
Lab3: defended
cf: done E
Lab4: started

+++ Term6/ml/labs/Lab3/valid.jl
@@ -1,3 +1,3 @@
-function accuracy(y::Vector{Float64}, y_true::Vector{Float64})
+function accuracy(y::Vector{Float64}, y_true::Vector{Float64})::Float64
     mapreduce((p, t) -> p * t > 0 ? 1 : 0, +, y, y_true) / size(y)[1]
 end
